(0723) refactor: #5 인증 - TokenDTO, TokenResponse 일원화(2102)
