+++ MAPICK/src/main/java/com/gitsunjaeab/mapick/api/auth/dto/response/LocalTokenResponse.java
@@ -1,10 +1,12 @@
 package com.gitsunjaeab.mapick.api.auth.dto.response;
 
+import com.gitsunjaeab.mapick.api.auth.dto.internal.TokenDTO;
 import com.gitsunjaeab.mapick.common.response.BaseApiResponse;
 import com.gitsunjaeab.mapick.common.response.ResponseCode;
-import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Data;
+
+import java.time.LocalDateTime;
 
 @Data
 @Builder
@@ -13,9 +15,9 @@ public class LocalTokenResponse implements BaseApiResponse {
     private String code;
     private String message;
     private LocalDateTime timestamp;
-    private TokenResponse token;
+    private TokenDTO token;
 
-    public static LocalTokenResponse of(TokenResponse tokenDTO) {
+    public static LocalTokenResponse of(TokenDTO tokenDTO) {
         return new LocalTokenResponse(
             ResponseCode.SIGNIN_SUCCESS.getCode(),
             ResponseCode.SIGNIN_SUCCESS.getMessage(),
