+++ MAPICK/src/main/java/com/gitsunjaeab/mapick/api/auth/dto/response/PasswordChangeResponse.java
@@ -1,5 +1,6 @@
 package com.gitsunjaeab.mapick.api.auth.dto.response;
 
+import com.gitsunjaeab.mapick.api.auth.dto.internal.TokenDTO;
 import com.gitsunjaeab.mapick.common.response.BaseApiResponse;
 import com.gitsunjaeab.mapick.common.response.ResponseCode;
 import lombok.Builder;
@@ -14,9 +15,9 @@ public class PasswordChangeResponse implements BaseApiResponse {
     private String code;
     private String message;
     private LocalDateTime timestamp;
-    private TokenResponse token;
+    private TokenDTO tokenDTO;
 
-    public static PasswordChangeResponse of(TokenResponse tokenDTO) {
+    public static PasswordChangeResponse of(TokenDTO tokenDTO) {
         return new PasswordChangeResponse(
             ResponseCode.CHANGE_PASSWORD_SUCCESS.getCode(),
             ResponseCode.CHANGE_PASSWORD_SUCCESS.getMessage(),
