+++ MAPICK/src/main/java/com/gitsunjaeab/mapick/infra/auth/token/JwtProvider.java
@@ -2,7 +2,7 @@ package com.gitsunjaeab.mapick.infra.auth.token;
 
 import com.gitsunjaeab.mapick.common.response.ResponseCode;
 import com.gitsunjaeab.mapick.domain.auth.Principal;
-import com.gitsunjaeab.mapick.domain.auth.TokenDTO;
+import com.gitsunjaeab.mapick.api.auth.dto.internal.TokenDTO;
 import com.gitsunjaeab.mapick.domain.member.Member;
 import com.gitsunjaeab.mapick.domain.member.MemberRepository;
 import com.gitsunjaeab.mapick.infra.auth.token.code.TokenType;
